requirements.txt
debugpy
django-cors-headers
django-materializecss-form
django-money
django
djangorestframework-simplejwt
djangorestframework
gunicorn
psycopg2-binary
pytest
social-auth-app-django
